Rename compaction module to compactors
